feat: added :style support
